--- .github/codeql/codeql-config.yml
@@ -1,22 +1,0 @@
-name: "CodeQL config"
-
-# Exclude the certificate-inspection helper from the rejectUnauthorized check.
-# fetchCertPemsFromUrl intentionally uses rejectUnauthorized: false because its
-# sole purpose is to retrieve TLS certificates from servers that may be
-# self-signed or expired. The SSRF denylist in fetchCertificateFromUrl ensures
-# only public IPs are reachable before this function is called.
-query-filters:
-  - exclude:
-      id: js/disabling-certificate-validation
-  # False positive on login endpoints. The query flags `if (!username || !password)`
-  # guarding a downstream `authenticateUser(username, password)` call because the
-  # credentials are user-controlled — but that is exactly how password auth works.
-  # The actual authorisation decision is made server-side inside authenticateUser().
-  # Pre-existing on main; flagged in PRs whenever the surrounding lines shift.
-  - exclude:
-      id: js/user-controlled-bypass
-
-# Belt-and-suspenders: also exclude the file from all queries so a CodeQL
-# database version change cannot re-surface the alert.
-paths-ignore:
-  - apps/web/lib/certificates/fetch.ts
